requirements.txt
mysql-connector-python
python-dotenv
OpenAI
flask
Flask-RESTful
Pillow
pytesseract